third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/worklet-import-blocked-expected.txt

CONSOLE ERROR: Refused to load the script 'http://127.0.0.1:8000/resources/worklet.js' because it violates the following Content Security Policy directive: "script-src 'unsafe-inline'". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback.

This is a testharness.js-based test.
Harness: the test ran to completion.
